# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the authenticated profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update the authenticated profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /businesses:
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Register a business account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Business registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/mine:
    get:
      operationId: getMyBusiness
      tags: [businesses]
      summary: Get the authenticated user's business
      responses:
        "200":
          description: Business registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Business not registered
  /businesses/{id}:
    get:
      operationId: getBusinessDetail
      tags: [marketplace]
      summary: Read a public business detail page
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Business detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessDetail"
        "404":
          description: Business not found
    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update an owned business
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /service-providers:
    post:
      operationId: createServiceProvider
      tags: [service-providers]
      summary: Register a service provider account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceProviderInput"
      responses:
        "201":
          description: Service provider registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProvider"
  /service-providers/mine:
    get:
      operationId: getMyServiceProvider
      tags: [service-providers]
      summary: Get the authenticated user's service provider profile
      responses:
        "200":
          description: Service provider registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProvider"
        "404":
          description: Service provider not registered
  /service-providers/{id}:
    put:
      operationId: updateServiceProvider
      tags: [service-providers]
      summary: Update an owned service provider
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceProviderUpdate"
      responses:
        "200":
          description: Updated service provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProvider"
...
  /admin/settings/new-on-shopnear:
    get:
      operationId: getNewOnShopNearSettings
      tags: [admin]
      summary: Read the new-listing window
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewOnShopNearSettings"
    put:
      operationId: updateNewOnShopNearSettings
      tags: [admin]
      summary: Update the new-listing window
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewOnShopNearSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewOnShopNearSettings"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: shopnear_session
  schemas:
...
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        profilePhoto:
          type: ["string", "null"]
          maxLength: 500
        city:
          type: ["string", "null"]
          maxLength: 80
        state:
          type: ["string", "null"]
          maxLength: 80
        address:
          type: ["string", "null"]
          maxLength: 250
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        preferredLanguage:
          type: ["string", "null"]
          maxLength: 10
        notificationsEnabled:
          type: boolean
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        accountType:
          type: string
          enum: [customer, business, service_provider, admin]
        profilePhoto:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        preferredLanguage:
          type: string
        notificationsEnabled:
          type: boolean
        locationAccuracy:
          type: ["integer", "null"]
        locationUpdatedAt:
          type: ["string", "null"]
          format: date-time
        locationEnabled:
          type: boolean
        locationPermissionStatus:
          type: string
        locationVisibility:
          type: string
          enum: [public, private]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, phone, email, accountType, profilePhoto, city, state, address, preferredLanguage, notificationsEnabled, status, createdAt]
    BusinessInput:
      type: object
      properties:
        businessName: { type: string, minLength: 2, maxLength: 160 }
        category: { type: string, minLength: 2, maxLength: 100 }
        businessLogo: { type: ["string", "null"], maxLength: 500 }
        coverPhoto: { type: ["string", "null"], maxLength: 500 }
        description: { type: ["string", "null"], maxLength: 2000 }
        workingHours: { type: ["object", "null"], additionalProperties: true }
        businessAddress: { type: ["string", "null"], maxLength: 250 }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        locationAccuracy: { type: ["integer", "null"] }
        locationUpdatedAt: { type: ["string", "null"], format: date-time }
        locationEnabled: { type: boolean }
        locationVisibility: { type: string, enum: [public, private] }
        whatsapp: { type: ["string", "null"], maxLength: 20 }
        phone: { type: ["string", "null"], maxLength: 20 }
        website: { type: ["string", "null"], maxLength: 500 }
        socialLinks: { type: ["object", "null"], additionalProperties: true }
      required: [businessName, category]
    BusinessUpdate:
      allOf:
        - $ref: "#/components/schemas/BusinessInput"
    Business:
      allOf:
        - $ref: "#/components/schemas/BusinessInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            ownerId: { type: string, format: uuid }
            verificationStatus: { type: string }
            approvedAt: { type: ["string", "null"], format: date-time }
            createdAt: { type: string, format: date-time }
            distanceKm: { type: ["number", "null"] }
            travelDistanceMeters: { type: ["number", "null"] }
            travelTimeSeconds: { type: ["number", "null"] }
          required: [id, ownerId, verificationStatus, approvedAt, createdAt]
    ServiceProviderInput:
      type: object
      properties:
        profession: { type: string, minLength: 2, maxLength: 120 }
        experience: { type: ["string", "null"], maxLength: 120 }
        skills: { type: array, items: { type: string }, maxItems: 20 }
        serviceRadius: { type: ["integer", "null"], minimum: 0, maximum: 500 }
        portfolioImages: { type: array, items: { type: string, maxLength: 500 }, maxItems: 12 }
        availability: { type: ["object", "null"], additionalProperties: true }
        phone: { type: ["string", "null"], maxLength: 20 }
        whatsapp: { type: ["string", "null"], maxLength: 20 }
        location: { type: ["string", "null"], maxLength: 250 }
        latitude: { type: ["number", "null"], minimum: -90, maximum: 90 }
        longitude: { type: ["number", "null"], minimum: -180, maximum: 180 }
        locationAccuracy: { type: ["integer", "null"], minimum: 0, maximum: 100000 }
        locationUpdatedAt: { type: ["string", "null"], format: date-time }
        locationEnabled: { type: boolean }
        locationVisibility: { type: string, enum: [public, private] }
      required: [profession]
    ServiceProviderUpdate:
      allOf:
        - $ref: "#/components/schemas/ServiceProviderInput"
    ServiceProvider:
      allOf:
        - $ref: "#/components/schemas/ServiceProviderInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            ownerId: { type: string, format: uuid }
            verificationStatus: { type: string }
            approvedAt: { type: ["string", "null"], format: date-time }
            createdAt: { type: string, format: date-time }
            distanceKm: { type: ["number", "null"] }
            travelDistanceMeters: { type: ["number", "null"] }
            travelTimeSeconds: { type: ["number", "null"] }
          required: [id, ownerId, verificationStatus, approvedAt, createdAt]
    NewOnShopNearSettings:
      type: object
      properties:
        days:
          type: integer
          minimum: 0
          maximum: 365
      required: [days]
...
    ProductInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: ["string", "null"], maxLength: 2000 }
        categoryId: { type: ["string", "null"], format: uuid }
        imagePaths: { type: array, items: { type: string, maxLength: 500 }, maxItems: 10 }
        primaryImagePath: { type: ["string", "null"], maxLength: 500 }
        priceCents: { type: integer, minimum: 0 }
        regularPriceCents: { type: ["integer", "null"], minimum: 0 }
        discountPriceCents: { type: ["integer", "null"], minimum: 0 }
        brand: { type: ["string", "null"], maxLength: 120 }
        condition: { type: ["string", "null"], maxLength: 80 }
        specifications: { type: ["object", "null"], additionalProperties: true }
        location: { type: ["string", "null"], maxLength: 250 }
        tags: { type: array, items: { type: string, maxLength: 40 }, maxItems: 30 }
        isAvailable: { type: boolean }
        isVisible: { type: boolean }
        status: { type: string, enum: [draft, published, hidden, scheduled, out_of_stock] }
        isFeatured: { type: boolean }
        scheduledAt: { type: ["string", "null"], format: date-time }
      required: [name, priceCents]
    Product:
      allOf:
        - $ref: "#/components/schemas/ProductInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            ownerId: { type: string, format: uuid }
            businessId: { type: string, format: uuid }
            status: { type: string }
            primaryImagePath: { type: ["string", "null"] }
            regularPriceCents: { type: ["integer", "null"] }
            discountPriceCents: { type: ["integer", "null"] }
            brand: { type: ["string", "null"] }
            condition: { type: ["string", "null"] }
            specifications: { type: ["object", "null"], additionalProperties: true }
            location: { type: ["string", "null"] }
            tags: { type: array, items: { type: string } }
            isFeatured: { type: boolean }
            scheduledAt: { type: ["string", "null"], format: date-time }
            publishedAt: { type: ["string", "null"], format: date-time }
            viewCount: { type: integer }
            favoriteCount: { type: integer }
            distanceKm: { type: ["number", "null"] }
            travelDistanceMeters: { type: ["number", "null"] }
            travelTimeSeconds: { type: ["number", "null"] }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
          required: [id, ownerId, businessId, status, createdAt, updatedAt]
    ServiceInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        description: { type: ["string", "null"], maxLength: 2000 }
        categoryId: { type: ["string", "null"], format: uuid }
        imagePaths: { type: array, items: { type: string, maxLength: 500 }, maxItems: 10 }
        primaryImagePath: { type: ["string", "null"], maxLength: 500 }
        priceFromCents: { type: ["integer", "null"], minimum: 0 }
        pricingOptions: { type: ["array", "null"], items: { type: object, additionalProperties: true }, maxItems: 20 }
        serviceRadius: { type: ["integer", "null"], minimum: 0, maximum: 1000 }
        availability: { type: ["object", "null"], additionalProperties: true }
        workingHours: { type: ["object", "null"], additionalProperties: true }
        emergencyService: { type: boolean }
        bookingReady: { type: boolean }
        estimatedDuration: { type: ["integer", "null"], minimum: 0 }
        location: { type: ["string", "null"], maxLength: 250 }
        tags: { type: array, items: { type: string, maxLength: 40 }, maxItems: 30 }
        isAvailable: { type: boolean }
        isVisible: { type: boolean }
        status: { type: string, enum: [draft, published, hidden, scheduled, out_of_stock] }
        isFeatured: { type: boolean }
        scheduledAt: { type: ["string", "null"], format: date-time }
      required: [name]
    Service:
      allOf:
        - $ref: "#/components/schemas/ServiceInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
            ownerId: { type: string, format: uuid }
            businessId: { type: ["string", "null"], format: uuid }
            providerId: { type: ["string", "null"], format: uuid }
            status: { type: string }
            primaryImagePath: { type: ["string", "null"] }
            pricingOptions: { type: ["array", "null"], items: { type: object, additionalProperties: true } }
            workingHours: { type: ["object", "null"], additionalProperties: true }
            emergencyService: { type: boolean }
            bookingReady: { type: boolean }
            estimatedDuration: { type: ["integer", "null"] }
            location: { type: ["string", "null"] }
            tags: { type: array, items: { type: string } }
            isFeatured: { type: boolean }
            scheduledAt: { type: ["string", "null"], format: date-time }
            publishedAt: { type: ["string", "null"], format: date-time }
            viewCount: { type: integer }
            favoriteCount: { type: integer }
            distanceKm: { type: ["number", "null"] }
            travelDistanceMeters: { type: ["number", "null"] }
            travelTimeSeconds: { type: ["number", "null"] }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
          required: [id, ownerId, businessId, providerId, status, createdAt, updatedAt]
...
    BusinessDetail:
      type: object
      properties:
        business: { $ref: "#/components/schemas/Business" }
        products: { type: array, items: { $ref: "#/components/schemas/Product" } }
        services: { type: array, items: { $ref: "#/components/schemas/Service" } }
        relatedBusinesses: { type: array, items: { $ref: "#/components/schemas/Business" } }
      required: [business, products, services, relatedBusinesses]
...